test(stella-core): split loop_escalation.rs's inline tests into a sibling file
driver/loop_escalation.rs sat 30 lines from the strict 1500-line
god-file ceiling, with ~900 of its 1470 lines being an inline
#[cfg(test)] mod tests block. Move the block verbatim into
driver/loop_escalation/tests.rs, the same sibling-tests.rs pattern
driver/tests.rs already uses for its own split-out submodules. Pure
test move, no production change.

Closes #3708
